ST SPEAr plgpio: adding support for gpio groups for interrupt registers
In Arch SPEAr13xx, plgpio gpio pins are grouped in number of four for
interrupt registers. Every group of four gpio's is represented by one bit
in IE and MIS registers. This patch adds support for this feature.
This will be reverted back once this feature is remove in next cuts of silicon

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>