ST SPEAr13xx: Check for parent clk select register before accessing it
Some clocks in SPEAr13xx may be part of parent clock set but at the same
time may not have proper selection register with them. This can happen
for virtual clocks and for some clock whose selection register belong to
the IP itself (rather than to the system configuration space). CLCD is
such an example.

Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>