AMBA-CLCD: Workaround: clcd clock must be enabled before configuring register
amba clcd driver assumes that the interface clock would always be enabled and
present and functional clock is seperate. On this assumption it disables
functional clock and continues with register configuration of clcd controller.
This fails for SPEAr as disabling clcd clock also disables its interface clock
and hence registers are not accessible.

Signed-off-by: Vipul Samar <vipulkumar.samar@st.com>