STMMAC : Support added for plat specific initilizations to be handled by driver.
To maintain the generic nature of the driver, and to be valid across all
synopsis ethernet GMAC cores, some part of the intializations for the core
is added into platform data which would differentiate between the GMAC cores

1. Added support for the Checksum offload engine types. The core may not have
the checksum engine or it may have checksum engine of type-1 or type-2

2. There are certain dependecy when the core is configured for RMII/SMII
interfaces, w.r.t the clocks needed. This is purely platform dependant.
A call back from setting the same hase been added.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>