SPEAr clock.c: vco_clk_set_rate must pass pll_clk to ddr_rate_acceptable
As vco and pll are closely bound to each other, we are passing pll's clk_rate
to vco_clk_set_rate routine. Which calls ddr_rate_acceptable with pointer to
vco's clk struct. It should instead pass pointer to pll's clock. This patch
finds the pll_clk, children to vco_clk, and passes it to ddr_rate_acceptable.

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>