SPEAr13xx: Add 2G/2G memory split to handle 1GB DDR
There are 2 possibilities to accommodate 1GB DDR in Kernel.
 - Enable high mem option in the kernel
 - Split virtual memory as 2G/2G for kernel and user space

It is observed that high mem option is unstable with ARMv7 SMP support.
Hence we enable 2G/@G memory split for now. This has an implication that
now the user space memory would be limited to 2GB of space, which was
earlier 3G with 1G/3G split.

Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>