usb/gadget/designware: wakeup proceedure aligned with documentation
Device user manual says about bit 0 (RES) of control register,
' To perform a remote wakeup resume the application sets this bit to 1,
then resets it to 0 after 1 ms. The UDC-AHB Subsystem signals the USB
host to resume the USB bus.'
The code was only setting this bit, changed it to reset the bit after 1
msec.

Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>