HID: Fix a crash in hid_report_raw_event() function.
Data Modul EasyTouch USB multitouch controller, issues a hid report with
a size equals to 0. The rsize value gets set to 536870912 and Linux is
crashing in the memset because the value is too big.
This patch tries to fix that by limiting rsize to HID_MAX_BUFFER_SIZE in
this case.

Signed-off-by: Armando Visconti <armando.visconti@st.com>