SPEAr13xx: Timer: Add support for saving timer registers during sleep.
The support has been provided for the clock source timer to save the control
register and stop the hardware timer before moving into sleep, then restore the
control register and start the hardware timer during wakeup.
The software requires some changes in future to add RTC support when the system
is in sleep, to account for sleep time, and then update the timer on wake up.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>