sound/asoc: Update audio support for SPEAr13xx
Change the name of the i2s driver to designware so that it can be used
across platforms.  Introduce support in the driver to handle capability
defined by platform on basis of which the driver would support either
record, play or both features. Also let platform define the maximum
number of channels to be supported by the driver.

Signed-off-by: Rajeev Kumar <rajeev-dlh.kumar@st.com>