stmmac/phy: Change the parameter of phy_reset from bsp_priv to *mii_bus
Currently stmmac is providing a hook phy_reset to the boards. This
callback is passing bsp_priv as an argument.  This argument must be
defined by the platform itself.

We require to access some phy resgisters in this callback and for that
we require mii bus instance. This seems to be more appropriate parameter
to be passed to this function as we can derive all instances from it
including mdio read/write functions.

Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>