mtd/nand/fsmc_nand: remove calls to mtd->suspend/resume
The fix solves the problem that causes the message:
"nand_resume called for a chip which is not in suspended state".

This happens because mtd->suspend and mtd->resume are already called
from the mtd layer, thus they were called a second time in fsmc_nand, so
mtd complains when trying to resume for the second time. The fix is to
remove call to mtd->resume/suspend.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@st.com>