usb/designware udc: Fix warning message when device not included
When designware usb device is not included by platform(board), gadget
drivers trying to register through usb_gadget_probe_driver fail with a
warning message as the device has not been initialized.

Include a check in the usb_gadget_probe_driver to check if device is
initialized to fix this.

Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>