SPEAr: PM : Add support for standby/suspend to ram.
The Power Management code for SPEAr13xx supports the standby and suspend to ram
modes. The system harware has power islands which are switched off during the
suspend to ram code execution.
The wake-up can be done using GPIO/Ethernet/USB/RTC/GPIO, when the system is in
suspend mode, with trigger routed through systems PMU. The wake up sources in
case of standby are routed through GIC.
The code makes use of PM notifiers to save the contents of VFP. Most of the
save/restore operations particular to SPEAr architecture are executed on the
"Always On Ram (AOR)".
Since the power islands are switched off during suspend to ram, the wake up code
control is passed from bootrom onto the AOR, for the restoration.

The Power Management code for SPEAr6xx/3xx supports suspend to ram mode, and
wake up could be done using GPIO/Ethernet/USB/RTC/GPIO.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>