SPEAr1340: Add gpio OR peripheral selection in padmux
In SPEAr1340 there are two levels of pad muxing
- pads as gpio OR peripherals
- If peripherals, then primary OR alternate peripheral.

The second level was already attended in padmux but first level was missed.
This patch adds supports for firstly making all pads as gpio first and then
enable whichever peripherals are required. For this earlier pmx structures are
updated and few more are added

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>