SPEAr1340: Support to enable pads as plgpio's
Pads can be configured as plgpios. It is difficult and complex to give separate
control structures for all pads in pmx, as we need to create three structures
for each pmx dev. So a single pmx_dev is created, which will update all pmx
registers, user just nedd to pass correct masks to update pads.

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>