SPEAr1340: PCIe: Modification as per changes in CR3 bit positions
LTSSM state was 5 bits in spear1300 and spear1310-A and use to occupy 4-0 CR3
app register. The modification in spear1340 and spear1310-B/C is that , the
LTSSM state of controller is 6 bits and thus occupy bit 5-0 of CR3 app register.
Because of this change all the upper bit position of CR_3 (from 6 onwards)
has been shifted by 1 position.

Signed-off-by: Pratyush Anand <pratyush.anand@st.com>