SPEAr1340: OTG: Temporary workaround : RXFifo size modification in GRXFSIZ
OTG driver does not have nice interface to program RX FIFO size. It
reads it from GRXFSIZ and again programs GNPTXFSIZ on the basis of this
read value.  However, in the case of dynamic Fifo it should be able to
be modified by each driver.In case of spear default values is 0x1770 and
system fails with some of the control transfer with this default values.
It should have some reasonable values like 0x400. So, in this patch
0x400 is hardcoded in the OTG GRXFSIZ register, so that driver can read
correct value.

Signed-off-by: Pratyush Anand <pratyush.anand@st.com>