sound/asoc/SPEAr: Handle rx data overrun interrupt properly
For receive/capture, two interrupt can be generated
   1. Data available and
   2. Data overrun interrupt

Data overrun interrupt is an error condition and it is generated when
incoming data is lost due to a full RX-FIFO. In case data overrun
interrupt is generated, s/w must disable I2S block and flush all the
RX-fifo.
To resume the functionality of audio capture it must enable I2S block at
the end of interrupt handler.

Signed-off-by: Rajeev Kumar <rajeev-dlh.kumar@st.com>