sound/asoc/spear: Disable unused receive channels
SPEAr1340 has upto 7.1 channel support. By default all the receive
channels are enabled. So before recording start we need to disable all
the rx channel. The number of channel will get enabled depending on the
mode(either 2.0, 3.1 audio or so on) selection. In case if all the
receives channel are enabled and we perform 2.0 capture, data loss will
be there.

Signed-off-by: Rajeev Kumar <rajeev-dlh.kumar@st.com>