asoc/spear: prevent race condition between tasklet and dma callback
Initially two DMA transfers are scheduled. After this each dma
completion schedules tasklet to start more DMA transfers. While the
tasklet (invoked from dma complete) executes, second DMA callback is
called and tasklet is rescheduled, while first was in the middle. When,
the tasklet resumes, it actually sees that dmacount is 0 and it has to
queue two transfers. Now, pending tasket is scheduled, and it sees that
dmacount is already two thus reporting the bug.

Correct this by not dropping lock in the tasklet.

Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>