sound/spear13xx_pcm: Add pcm_running to know current status of transfer
When pcm's trigger routine with STOP command is called we simply clear all DMA
channels and their queues. While doing that it might be possible that we are
already in middle of scheduling new transfers.

So, pcm_running is introduced to check we should continue with scheduling new
transfers or we should STOP.

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>