media: soc-camera: value of 'VS6725_DEVICE_ID_HI' macro changed to 0x00
The VS6725 user manual states that the value of LOWER byte of DEVICE_ID should
be 0x002, however the value read from the actual sample on the SPEAr1340 CPU
card is 0x00.

This mismatch causes the vs6725 probe to fail. While raising this query to the
VS6725 designers, changed this value in the driver to allow further tests on the
VS6725 sensor driver.

This patch solves the bug "[Issue #142674] Camera sensor device id mismatch"

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>