st gyroscope: Fix to display correct gyro settings
The gyroscope device's polling rate and full scale range can be set by the user.
There are fixed values of polling rate and full scale range that the device can
take. Both of these can be set and displayed thro' the device attributes exposed
by the driver. Currently, the driver reads static data to display these
parameters. As a result, even if the user updates any of these, obsolete value
is returned.

This patch addresses two bugs which are similar in nature by defining a new
structure to hold the current paramters for the device. The two bugs are:
1. Issue # 124374 - 'range' parameter is not set properly
2. Issue # 124376 - 'pollrate_ms' parameter setting issue

Signed-off-by: Amit Virdi <amit.virdi@st.com>