dma: dw_dmac: value of 'NR_DESCS_PER_CHANNEL' macro changed to 2048
The NR_DESCS_PER_CHANNEL macro value proved insufficient to handle the
requirements of the VS6725 camera sensor connected on the SPEAr1340
board which can support 1600*1200 resolution.

Although is is not a very elegant solution as there must be a way
somehow to configure this value through some platform data (thus
different for each peripheral using the system DMA), future patches will
try to address this issue.

At the moment we only modify this macro to enable further testing on
VS6725 camera sensor driver.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>
Reviewed-by: Viresh Kumar <viresh.kumar@st.com>