dwc/otg: correct implementation of ep disable and flush fifo
There are some defined steps to disable an endpoint.
It can not be disabled just by setting epdisable bit.
Now, it is disabled as specification says.

During stall of and endpoint, first stall bit should be
set and then endpoint should be disabled.

In dwc_otg_pcd_ep_disable, TX fifo must be flushed along with
disabling of endpoint.

In DMA mode if Fifo is flushed only once then, DMA will write
data into it again. So, it must be flushed till DMA transfer is
complete.

Signed-off-by: Pratyush Anand <pratyush.anand@st.com>