dmaengine/dw_dmac: No need to disable interrupts in dw_dma_off
In case of suspend, we mask all interrupts from dw_dma_off() routine, but in
resume we don't restore their values. We unmask them only from
dwc_alloc_chan_resources() routine, and that is not called again after resume.
So, dma transfers do not start on channels for which channels were allocated
before suspend.

To solve this, we could have saved state of these registers, but actually there
is no need of doing that too. Why disable interrupts when we have disabled
complete peripheral earlier. So, its better not to mask these interrupts on
suspend.

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>