media: soc-camera/spear_camif: DMA channel requests moved to 'init_videobuf'
This patch addresses the bug, Issue #141516, for CAMIF driver.

Originally, the DMA channel requests were made in the _probe_ routine
to be in sync with implementation of similar drivers already present in
mainline.

However, this caused the DMA channels to be acquired through-out and other
peripherals requesting the system DMA resources can be blocked in-case
one or more channels are acquired by CAMIF, even when no user-space application
is running on top of the same.

Accordingly, the DMA channel requests have now been moved to
'camif_init_videobuf' function which is called by the soc-camera core when
the user space application calls the 'open' ioctl for this node. The DMA
channels are released in 'camif_remove' routine which is called by the
soc-camera core when the user-space application calls the 'close' ioctl.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>