SPEAr/padmux: Allow padmux to be reconfigured multiple times
Till now, padmux can be configured only once using call to pmx_register routine.
With addition of plug boards of SPEAr1340, it is now required to configure pmx
again from plug boards.

This patch removes static declaration of pmx_devs_enable() routine and adds it
prototype in padmux.h file. After this padmux can be reconfigured as many times
as required with a call to pmx_register() for this first time pmx is configured,
and call to pmx_devs_enable() every time pmx is reconfigured.

Also, kfree(pmx) is removed so that pmx can be reused again.

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>