SPEAr/spi: Repair broken DMA support for spi
SPI was believed to be working with DMA, but actually it wasn't. There are two
places where we need to enable DMA support,
- platform_data of amba_device for spi
- controller_data of spi_devices

First was always taken care of, but second was never done. So although we got
prints that DMA channels are acquired, DMA never worked. It was always Interrupt
mode.

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>