SPEAr/clock: Mark all VCOs used in system clock as ENABLED_ON_INIT
Mark all VCOs (vco1, vco2 and vco3) which can be used a source for
system clocks (feeding cpu and amba) as ENABLED_ON_INIT. This is because
if we don't do it then it can be potentially turned off if nobody uses
it.

ENABLED_ON_INIT would ensure that we enable it on bootup and hence would
always be marked as in use. VCO1 is already marked as ENABLED_ON_INIT.

Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>