SPEAr320/DMA: Change DMA master to MASTER 2 for RAS peripherals
There are two masters present in DMAC controller and both masters have access to
specific address range. For RAS IP's master 2 is fixed in RTL and thus we must
configure DMA channels with Master 2 for transfers with RAS peripherals.

Currently, master 1 is configured for RAS peripherals. This patch configures
master 2 for RAS peripherals.

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>