SPEAr/clock: fix to maintain clock hierarchy on clk_reparent
Currently, when we call clk_set_parent, debugfs entry for clk is deleted
from previous, and then is created inside new parent. When this is done,
state of children clocks is lost when we delete current clocks dentry.
And due to this few clocks directly appear in /sys/kernel/debugfs
instead of their actual hierarchical position.

This patch fixes this issue by using debugfs_rename routine instead.
With this state of children is not lost and they remain in the correct
hierarchy in debugfs.

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>