SPEAr13xx: PCIe: Add support for Suspend Resume
This patch privides in the support for the Suspend and resume functionality
for the PCIe host controller.
There is a limitation related to not plugging the suspend/resume call backs to
the platform bus. This has been avoided, as during the resume operations, PCIe
bus devices were trying to access the PCIe space before the host being enabled.
This is so because the platform bus was being resumed after the PCIe bus.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>