cpufreq: change to slow mode if required
For all changes in cpu frequency involving pll we must adhere to
following procedure.
   - switch system to 'slow' mode, so that cpu can run on alternate
     (crystal) frequency as pll would be turned off during rate change.
   - now switch pll rate
   - change system mode back to 'normal' to drive cpu from pll

This is not applicable only in case of spear1340 when system synthesizer
is feeding cpu, as synthesizer can be re-programmed on runtime.

SPEAr3xx/6xx are also exception as DDR is also running on PLL1 hence the
code changing the rate pf PLL must not be run from DDR. This has already
been handled in the clock framework itself.

Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>