SPEAr: configure proper value of programmable burst length (pbl)
It is observed that devices with older version of gmac(v3.10a) are
limited  to incr 32 burst where as this can go up to incr 128 in case of
SPEAr1340.

Further we need to take into account that PBL4 (pbl * 4) is actually
PBL8 (pbl * 8) in v3.61a (belonging to SPEAr1340).

Hence,
   - set PBL4 option
   - program pbl to 8 (for incr 32 burst) for older version
   - program pbl to 16(for incr 128 burst/beat) for newer version

Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>
Signed-off-by: Vikas Manocha <vikas.manocha@st.com>