SPEAr13xx PM: Common wake code for SPEAr13xx/1340.
The SPEAR13xx platform Power Management Hardware is not cent percent backward
compatible with its previous architecture. The code integration in the same low
level files causes a lot of clutter, and more importantly the code has to be
executed from always on ram, where we have the memory size limitations.
Since the wake up code is common across the platforms, this patch separates the
wake up code for the SPEAr13xx/1340 platforms. This patch uses a scratch pad
location reserved on the Always on RAM, to identify the correct resume location.
The scratch pad location is filled up by the respective platforms before moving
into sleep mode.
This patch also adds the Support for runtime check to Suspend the apt platform.
This patch puts the system into suspend based on the architecture checks done
on the run time.The Makefile is also realigned to support the code changes.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>