media: soc-camera: Fix vs6725 register read and write routines
There was a issue in handling of the I2C offsets for VS6725 sensor.
The buggy code considered and implemented offsets as 8-bit, whereas
VS6725 sensor offsets are 16 bits.

As a result data read from and written to VS6725 via I2C interface was
not getting reflected properly.

This patch fixes the same and accordingly the '.len' field in the I2C register
read and write messages have been modified.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>