SPEAr1340/evb: Removing support for on-board camera sensor
As confirmed by HW designers, it is difficult to mount a camera sensor on
SPEAr1340 CPU card.

So, although a slot has been provided on the SPEAr1340 CPU card to mount a
camera sensor, we will have to use the camera plug boards to have access to a
camera sensor.

This patch removes the support of camera sensor in SPEAr1340 evb file to
reflect the same.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>
Acked-by: Viresh Kumar <viresh.kumar@st.com>