SPEAr1340/plugboards: Camera plug board support updated to support single sensor
Current version of Camera Plug board which can be mounted on SPEAr1340,
supports only a single camera sensor and although slots are provided to house 4
camera sensors, all the sensors cannot be mounted together on the plug board.

The reason for the same is that all camera sensors have a I2C address of 0x20,
and currently no on-board mechanism has been provided to differentiate I2C
transfer between the four sensors.

As discussed with Franco, a new plug board will be designed to support all four
camera sensors at the same time. Till then we have support only for a single
sensor on the plug-board.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>
Acked-by: Viresh Kumar <viresh.kumar@st.com>