net/stmmac: Resolve WOL conflict between ethtool and sysfs
Some of the platforms require an explicit configuration of the wake up source
interrupts. The kernel code specifically related to ARM requires an explicit
configuration of these sources at VIC, which is not being catered using the
sysfs interface. The ethtool while setting the wol option tries to register the
wol device and then enable wol interrupt. In case the sysfs has already
registered the wol device, this patch adjusts for that registration and
configures the wake-up interrupt.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>