char/spear_adc: Fix suspend/resume issues
Function calling specially in suspend/resume isn't handled properly. Return type
of routines are not at all checked and their sequence is also not the best.

Similar issues are also present in probe routine.
Another routine for configuring default configuration is also created, so that
it can be called from multiple places.

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>