media: soc-camera: Do not stop subdev streaming when camera is detached
Due to the peculiar nature of the test subdevice used with SPEAr
CAMIF interface, we should not stop the subdev streaming when Camera
is detached.

If the streaming is stopped on the VS6725 camera sensor subdevice,
it requires a long sequence of patch regsiters again to be written, to start
functioning again.

Instead relegate the responsibility of calling stream-off on the
subdev to the application space.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>