media: soc-camera: Fix HSYNC, VSYNC polarity for VS6725 (incompatible with VESA)
The HSYNC and VSYNC polarity terminology defined in VS6725 User Manual is
incompatible with VESA standards. A number of host interfaces (like
SPEAr CAMIF) use the VSYNC, HSYNC ploarity terminology as defined in VESA.

So, this patch reverses the HSYNC and VSYNC plority notions for the sensor,
to be compatible to the externel world.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>