Video/display: Standard CEA format definition correction
EDID information provided by every sink device has information about
display format. This information is provided as a video code which is
part of CEA standard.
Information for some of the video code were correct while for some of
them it was not correct. This patch corrects information as well as the
way this information was presented.
Only few formats have been added. Rest of other can also be added
subsequently.
It is believed that, these information should be present in a common
place where every video device can use it.

Signed-off-by: Pratyush Anand <pratyush.anand@st.com>