media: soc-camera: Fix half-frame capture at start in SPEAr CAMIF
The 'frame_start_frame_end' ISR was giving higher priority to
frame-start interrupt, in case both frame-start and frame-end
events are reflected in the interrupt status register.

This patch corrects this behavior by ensuring that we always
do valid DMA activity at a frame-end which should occur only
after a valid frame-start event.

In this way, we remove the possibility of capturing 1st frame as
a half-frame.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>