media: soc-camera: Modify 'end-of-array' marker for vs6725 patches
As recently discovered from the VS6725 evaluation scripts provided by ST,
a write of 0xff is valid at sensor patch-ram address 0xffff.
So, {0xffff, 0xff} cannot be used as an 'end-of-array' marker for vs6725
patches.

This patch introduces {0x0000, 0x00} as a new 'end-of-array'
marker as the register 0x0000 is READ-ONLY and hence can never
be written-to.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>