SPEAr13xx PM: Deactivate MPMC reset during resume
The DDR controller has been protected off the spikes that may be generated
during the reset triggered by suspend to ram. This protection has been
provided by using GPIO control of DDR clk enable and reset.
During the resume operations, this patch deactivates the reset control in the
software; and adds in some delay to ensure DDR is out of refresh before the
access is granted.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>