SPEAr320: c_can: Provide a way to send quirks specific to c_can from platform
C_CAN controller modules have some CLK HW issues in SPEAr320 which have
been fixed in SPEAr320s silicon.

Accordingly a mechanism needs to be added where-in whether the SW quirks
are to be applied or not need to be determined as per the SoC ID information.

This patch adds this support.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>