net: can: Add fixes for SPEAr320 silicon issue in c_can controller
SPEAr320 has a silicon issue in the way CAN clk is used inside the
c_can conroller. Details of this issue can be seen in the ERRATA document,
available here:
http://www.st.com/internet/mcu/product/247244.jsp#ERRATA_SHEETS

As a result a series of SW fixes need to be added to provide a workaround
for these issues. These SW fixes are enabled only when the SoC ID read from
the SOC_CORE_ID register (available in MISC space) is for SPEAr320.

The following fixes have also been added to make the SW wokarounds work:
1. The receive object numbers which are re-activated in
   'c_can_activate_all_lower_rx_msg_obj' was incorrect, as FIRST
   receive object was never re-initailized.

2. In 'c_can_do_rx_poll' the algo to receive message object was incorrect.
   As a result the received messages were out-of-order and at times some
   data was also lost.

This patch has been tested on SPEAr320 PLC board with SPEAr320 and SPEAr320s
silicons.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>