mfd/stmpe: Pass partnum as param to stmpe_probe()
partnum is required during probe to get variant's info. Currently partnum was
getting set in i2c interface after it is used and is not at all getting set in
spi interface.

This can be passed as param to stmpe_probe, so that it is available early. With
this, there is no need of i2c interface's init() routine, as it does nothing
else than calling dev_set_drvdata(). So better remove this routine for i2c.

partnum defined in struct stmpe is of no use now, as it is only used once during
probe of stmpe, so we can remove it from struct stmpe too.

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>