SPEAr/plgpio: Return true/false from is_plgpio_set()
is_plgpio_set() routine is used to get value of plgpio pins. Currently this
routine returns:

value of reg & (1 << offset of bit)

With this, if we are using pins like 31, 63, etc (Most significant bit set), and
we get these pins as SET, we return 0x80000000 from gpio_get_value. Now, this
routine returns -ve on errors and our return value is considered to be an error.
For this fix is_plgpio_set() routine to return true/false only.

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>