SPEAr13xx: PM : Add support to DDR Phy Shutoff.
This patch adds to PM the support to DDR Phy Shutoff. This option is
enabled setting ddr_phy_no_shutoff to 0 in ST's SPEAr13xx PCM_CFG
misc register. The DDR Phy Shutoff selects the SPEAr DDR pads,
DDRIO_VDD1V8_1V5_OFF and DDRIO_VDD1V2_OFF, as sources to control the
lines for the switching of the DDRPHY to the external power supply.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@st.com>