media: soc-camera: Allow VS6725 to be powered-off/on by SoC-camera framework
The SoC camera framework provides provision to turn ON/OFF a camera sensor when
the open/close of 'soc_camera' driver is called by the user-space application.

This requires that the default patches for the camera sensor are written
when the camera STREAMON is requested by the user-space application and
not from the _probe_ routine.

In addition the delay value required after the patches are written needs
to be altered to ensure that the sensor never outputs 1st frame as
a BLACK one.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>