media: soc-camera: Fix 'succesive application launch not working' issue for CAMIF
This patch addresses the Issue #150433 raised in CODEX for SPEAr CAMIF.

With earlier version of CAMIF driver succesive launchs of test application
were not working. The driver worked fine (captured frames) for the very 1st
time, but successive test application launchs returned 'select_timeout' error.

This was happening due to incorrect management of DMA descriptors in the
CAMIF driver, which was causing some DMA descriptors to remain unaccounted
for when a 'close' was being called from the user-space application.
Successive 'open' requests from application caused crash in DMA driver.

This patch completely changes the way 'tx_submit' is called
for the videobufs being queued by the test application:
	1. DMA descriptors are prepared and submitted for each
	   videobuf being queued in the _videobuf_queue ioctl itself.
	2. The state of each submitted videobuf is kept initially as
	  VIDEOBUF_QUEUED.
	3. The CAMIF interrupts are kept DISABLED initially and only HSYNC/
	   VSYNC interrupts are enabled in _videobuf_queue ioctl.
	4. When the 1st frame-end interrupt is received we mark the state of
	   1st videobuf present in the queue (which was filled in
	   _videobuf_queue), as ACTIVE and unmask DMA requests.
	5. In the DMA completion callback, we simply mark the videobuf
	   as DONE and extract the next videobuf from the queue and mark it
	   as ACTIVE.

This patch also reworks the interrupt configuration routines to ensure
correct behaviour.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>