SPEAr13XX: Provide a more appropriate name for clk, irq, pmx of Video Input Port
The Video Input Parallel port is referred throughout the SPEAr1340 user manual
and other documents as VIP.

So, this patch changes the name of the VIP clk, to 'spear_vip' which is also in
conformance with the VIP driver name.

Similarly the names for VIP irq(s), pmx(s), .. have been changed to 'vip'.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>
Reviewed-by: Viresh Kumar <viresh.kumar@st.com>