USB: ehci_spear: fix controller wake up setting during resume
The host controller suspend operations are coupled with the host
controller power domain being moved into off state. On resume operation
of Power Domain into on state, the host controller releases the owner
ship of the port. By default and EHCI enumerated device is then
re-enumerated as an OHCI device in all subsequent suspend and resume
operations.

To fix this issue, a workaround has been added to enable the port power
and regain the ownership of the port for EHCI controller.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>
Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>