net: can: Provide a more generic interface to configure c-can message objects
Depending on the underlying platform (in this case SPEAr320/320S), the
configuration of the c-can message objects can change. For e.g. due
to a silicon bug in SPEAr320, there can be only one TX message object,
whereas the TX message objects can be greater than 1 for SPEAr320S.

This patch adds this support.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>