SPEAr1340 evb: Switch off/on SATA power domain in suspend
The SATA power domain has been explicitly switched off and on during the
suspend and resume processes. This fix has been added in observation of
a SATA behaviour during the resume processing.

The resume processing had the necessary processing related to
initializing the controller and resuming the host. There were problems
related to SATA command processing even after the link was getting up.
The commands were getting timed out. This patch brings in consistency
w.r.t the states expected when the controller is reinitialized by
switching off and on the SATA power domain.

Signed-off-by: Deepak Sikri <deepak.sikri@st.com>