SPEAr320: Don't register shared irq ras3 for 320s
In SPEAr320s, shared irq ras3 is shared between plgpio and I2S. Since I2S
interrupts are required to be enabled at controller level (required for DMA to
work), and there is no mechanism to mask it in shirq level, we have to disable
interrupt on VIC. For this we must not register ras3 shirq. With this PLGPIO
interrupts also go away.

Signed-off-by: Viresh Kumar <viresh.kumar@st.com>