SPEAr1340/plugboard: Remove support for CAM0 plug-board camera sensor
With SPEAr1340 Rev2.1 board being available which has an on-board camera
module connected to CAM3 interface (numbering of the camera interfaces
is CAM0, 1 and so on..), the CAM0 plug-board camera sensor can no longer
be supported as all the camera sensors have an I2C address of 0x20 (in
7-bit addressing convention this becomes 0x10) and are connected to I2C.0
controller.

This patch removes the support of CAM0 plug-board camera sensor from
spear1340_plugboards.c file.

Signed-off-by: Bhupesh Sharma <bhupesh.sharma@st.com>