sound/asoc: Make audio driver generic according to dmaengine.
The changes are required because we need to keep same driver for
SPEAr13XX and SPEAr320s. The argument passing for configuring dmaengine
are different in both cases. So to make audio driver generic, we need to
pass dma parameters in a fixed pattern to dmaengine.

The patch also contains changes related to spear320s audio driver.

Reviewed-by: Viresh Kumar <viresh.kumar@st.com>
Signed-off-by: Rajeev Kumar <rajeev-dlh.kumar@st.com>