SPEAr1340: Add SPEAr1340 i2c bus recovery routine
This patch adds the ST's SPEAr1340 specific implementations of the
DesignWare I2C bus recovery procedure after timeout. This routine would
be a part of platform data for DesignWare I2C controller and shall be
called when I2C bus hangs.

The idea is to take control of I2C clk pads as programmable gpio and
then toggle it in order to recover the bus.

Signed-off-by: Vincenzo Frascino <vincenzo.frascino@st.com>
Signed-off-by: Shiraz Hashim <shiraz.hashim@st.com>